[refactor] SearchInput 클래스를 상속받는 AutoComplete, RecentSearch 클래스 생성 - 기존에 SearchInput의 메서드를 용도에 따라 각각의 하위 메서드로 넣어주었다. - toggleInputFocusClass과 resetInputText 메서드는 인스턴스 선언없이 유틸로 사용하기 위해 static 키워드를 붙여주었다.
